Updated API and refactor
