+++ scanomatic/ui_server_data/js/src/api/index.js
@@ -1,5 +1,8 @@
 import $ from 'jquery';
-import Duration from './Duration';
+import Duration from '../Duration';
+import API from './API';
+
+export { default as getPlateGrowthData } from './getPlateGrowthData';
 
 const GetSliceImagePath = '/api/calibration/#0#/image/#1#/slice/get/#2#';
 const GetTranposedMarkerPath = '/api/data/fixture/calculate/';
@@ -8,52 +11,6 @@ const GetGrayScaleAnalysisPath = '/api/data/grayscale/image/';
 const secondsPerMinute = 60;
 const secondsPerHour = 3600;
 const secondsPerDay = 86400;
-
-class API {
-    static get(url) {
-        return new Promise((resolve, reject) => $.ajax({
-            url,
-            type: 'GET',
-            success: resolve,
-            error: jqXHR => reject(JSON.parse(jqXHR.responseText).reason),
-        }));
-    }
-
-    static delete(url) {
-        return new Promise((resolve, reject) => $.ajax({
-            url,
-            type: 'DELETE',
-            success: resolve,
-            error: jqXHR => reject(JSON.parse(jqXHR.responseText).reason),
-        }));
-    }
-
-    static postFormData(url, formData) {
-        return new Promise((resolve, reject) => $.ajax({
-            url,
-            type: 'POST',
-            contentType: false,
-            enctype: 'multipart/form-data',
-            data: formData,
-            processData: false,
-            success: resolve,
-            error: jqXHR => reject(JSON.parse(jqXHR.responseText).reason),
-        }));
-    }
-
-    static postJSON(url, json) {
-        return new Promise((resolve, reject) => $.ajax({
-            url,
-            type: 'POST',
-            data: JSON.stringify(json),
-            contentType: 'application/json',
-        })
-            .then(
-                resolve,
-                jqXHR => reject(JSON.parse(jqXHR.responseText).reason),
-            ));
-    }
-}
 
 export function GetSliceImageURL(cccId, imageId, slice) {
     const path = GetSliceImagePath.replace('#0#', cccId).replace('#1#', imageId).replace('#2#', slice);
